+++ src/Phan/Language/Type/MixedType.php
@@ -54,6 +54,11 @@ final class MixedType extends NativeType
         return ArrayType::instance(false);
     }
 
+    public function isArrayOrArrayAccessSubType(CodeBase $unused_code_base) : bool
+    {
+        return true;
+    }
+
     public function isPrintableScalar() : bool
     {
         return true;  // It's possible.